requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.15.0
google-auth==1.10.0
google-cloud==0.34.0
google-cloud-core==1.1.0
google-cloud-logging==1.14.0
google-cloud-storage==1.24.1
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
numpy==1.18.0
pandas==0.25.3
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
twilio==6.35.1
urllib3==1.25.7
